move test case to test_sort_and_select.py & fine-turning in doc
